# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Portuguese Autónomo & Tax Optimization API",
    "version": "2.0.0",
    "description": "Local-first operations & accounting assistant for Portuguese Autónomos. Two layers: (1) stateless CIRS/CRCSPSS/CIVA tax & Social Security simulators — Recibos Verdes, e-Fatura deductions, Segurança Social quarterly variation (Art. 163º CRCSPSS), and Unipessoal Lda break-even comparison; (2) a persistent profile + transaction ledger that auto-derives a personalized compliance obligations timeline and dashboard, so numbers don't have to be re-entered by hand on every call. This is a local system of record and calculator — it does not file anything with the AT or Segurança Social on your behalf.",
    "contact": {
      "name": "Portuguese Autónomos Open Source Community",
      "url": "https://github.com/n4ouri/pt-autonomo-api"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
    "/api/v1/simulate/regime-simplificado": {
      "post": {
        "summary": "Simulate CIRS Regime Simplificado & 15% Expense Justification",
        "description": "Calculates taxable income under CIRS Art. 31, applies Start-of-Activity discounts (50% yr 1, 25% yr 2), checks mandatory 15% expense justification, and calculates progressive IRS.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "annualGrossServices": {
                    "type": "number",
                    "example": 60000
                  },
                  "annualGrossSales": {
                    "type": "number",
                    "example": 0
                  },
                  "annualGrossOtherServices": {
                    "type": "number",
                    "example": 0
                  },
                  "activityYear": {
                    "type": "integer",
                    "example": 3
                  },
                  "annualSSPaid": {
                    "type": "number",
                    "example": 7200
                  },
                  "businessExpenses": {
                    "type": "number",
                    "example": 2500
                  },
                  "homeOfficeEligibleExpenses": {
                    "type": "number",
                    "example": 2400
                  },
                  "withheldTax": {
                    "type": "number",
                    "example": 12000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Simulation completed successfully"
          }
        }
      }
    },
    "/api/v1/simulate/ss-quarterly": {
      "post": {
        "summary": "Simulate Segurança Social Quarterly Base & -25%/+25% Variation",
        "description": "Calculates relevant quarterly income (70% services / 20% sales), monthly base (capped at 12 IAS), and strategic -25% to +25% escalão variations under Art. 163 CRCSPSS.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quarterlyGrossServices": {
                    "type": "number",
                    "example": 15000
                  },
                  "quarterlyGrossSales": {
                    "type": "number",
                    "example": 0
                  },
                  "isFirstYearExempt": {
                    "type": "boolean",
                    "example": false
                  },
                  "hasConcurrentEmployment": {
                    "type": "boolean",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Quarterly SS calculation result"
          }
        }
      }
    },
    "/api/v1/simulate/company-vs-recibos": {
      "post": {
        "summary": "Break-even Simulator: Sociedade Unipessoal Lda vs Recibos Verdes",
        "description": "Compares effective tax rates, accounting costs, meal allowance savings, and net cash between Recibos Verdes (CIRS) and Unipessoal Lda (IRC 17% + MOE salary).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "annualGrossRevenue": {
                    "type": "number",
                    "example": 65000
                  },
                  "annualOperationalExpenses": {
                    "type": "number",
                    "example": 6000
                  },
                  "monthlyDirectorSalary": {
                    "type": "number",
                    "example": 1400
                  },
                  "monthlyAccountingFee": {
                    "type": "number",
                    "example": 150
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Financial comparison and recommendation"
          }
        }
      }
    },
    "/api/v1/efatura/optimize": {
      "post": {
        "summary": "Optimize e-Fatura Tax Deductions & Ceiling Headroom",
        "description": "Evaluates statutory deduction caps for Saúde, Educação, Imóveis, Despesas Gerais, Benefício IVA (restauração, passes, oficinas), and PPR.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "categories": {
                    "type": "object",
                    "properties": {
                      "despesasGerais": {
                        "type": "number",
                        "example": 600
                      },
                      "saude": {
                        "type": "number",
                        "example": 800
                      },
                      "educacao": {
                        "type": "number",
                        "example": 300
                      },
                      "habitacao": {
                        "type": "number",
                        "example": 2400
                      },
                      "ivaRestauracao": {
                        "type": "number",
                        "example": 1200
                      }
                    }
                  },
                  "pprInvested": {
                    "type": "number",
                    "example": 1500
                  },
                  "age": {
                    "type": "integer",
                    "example": 32
                  },
                  "isMarried": {
                    "type": "boolean",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deduction optimization report"
          }
        }
      }
    },
    "/api/v1/vies/check": {
      "post": {
        "summary": "VIES & VAT Cross-Border Compliance Rules",
        "description": "Determines mandatory invoice mention (Reverse charge Art. 6 n. 6 al. a CIVA), VAT rates, and VIES recapitulativa declaration deadlines.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "clientType": {
                    "type": "string",
                    "enum": [
                      "PT",
                      "EU_B2B",
                      "EU_B2C",
                      "NON_EU"
                    ],
                    "example": "EU_B2B"
                  },
                  "invoiceAmount": {
                    "type": "number",
                    "example": 4500
                  },
                  "hasValidVIES": {
                    "type": "boolean",
                    "example": true
                  },
                  "isArt53Exempt": {
                    "type": "boolean",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "VIES compliance instructions and invoice clause"
          }
        }
      }
    },
...
    "/api/v1/simulate/nhr": {
      "post": {
        "summary": "Simulate NHR 20% Flat Rate & IFICI Tax Benefit",
        "description": "Calculates tax under NHR (Residente Não Habitual / IFICI) flat 20% rate for high added-value professional activities (Portaria n.º 12/2010).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "annualGrossServices": {
                    "type": "number",
                    "example": 75000
                  },
                  "annualSSPaid": {
                    "type": "number",
                    "example": 7500
                  },
                  "businessExpenses": {
                    "type": "number",
                    "example": 3750
                  },
                  "withheldTax": {
                    "type": "number",
                    "example": 15000
                  },
                  "isEligibleHighValueActivity": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "NHR simulation results and tax comparison"
          }
        }
      }
    },
    "/api/v1/simulate/tax-free-allowances": {
      "post": {
        "summary": "Simulate Tax-Free Allowances (Km, Meal Card, Per Diems)",
        "description": "Calculates annual net tax-free cash extraction potential via Km allowances (0.40€/km), meal card (10.20€/day), and travel per diems with 0% IRS and 0% SS.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "monthlyKmDriven": {
                    "type": "number",
                    "example": 1000
                  },
                  "workDaysPerMonth": {
                    "type": "integer",
                    "example": 22
                  },
                  "nationalTravelDaysPerYear": {
                    "type": "integer",
                    "example": 10
                  },
                  "foreignTravelDaysPerYear": {
                    "type": "integer",
                    "example": 5
                  },
                  "isCompanyManagingPartner": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tax-free extraction summary and legal breakdown"
          }
        }
      }
    },
    "/api/v1/simulate/ppc": {
      "post": {
        "summary": "Forecast Pagamentos por Conta (PPC) & Suspension Eligibility",
        "description": "Calculates July, September, and December PPC tax instalments (Art. 102.º CIRS) and evaluates eligibility to legally suspend the 3rd payment (Art. 102.º n.º 7).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "priorYearNetTax": {
                    "type": "number",
                    "example": 12000
                  },
                  "priorYearWithholding": {
                    "type": "number",
                    "example": 0
                  },
                  "currentYearEstimatedTax": {
                    "type": "number",
                    "example": 8000
                  },
                  "currentYearWithholding": {
                    "type": "number",
                    "example": 2000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PPC forecast and suspension analysis"
          }
        }
      }
    },
    "/api/v1/simulate/irs-jovem": {
      "post": {
        "summary": "Simulate IRS Jovem Exemption (Artigo 12.º-B do CIRS)",
        "description": "Calculates tax exemption under IRS Jovem (up to 100% year 1, capped at 55x IAS) for young professionals up to 35 years old.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "annualTaxableIncome": {
                    "type": "number",
                    "example": 30000
                  },
                  "age": {
                    "type": "integer",
                    "example": 27
                  },
                  "yearOfBenefit": {
                    "type": "integer",
                    "example": 1
                  },
                  "educationLevel": {
                    "type": "integer",
                    "example": 6
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "IRS Jovem simulation results"
          }
        }
      }
    },
    "/api/v1/simulate/ss-benefits": {
      "post": {
        "summary": "Calculate Social Security Benefits (Parentalidade, Baixa Médica, Desemprego)",
        "description": "Computes maternity/paternity 100% subsidy, sickness benefit after 10-day waiting period, and unemployment eligibility for independent workers.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "monthlyContributionBase": {
                    "type": "number",
                    "example": 3000
                  },
                  "registeredMonthsInLastYear": {
                    "type": "integer",
                    "example": 12
                  },
                  "economicDependencePercentage": {
                    "type": "number",
                    "example": 60
                  },
                  "hasDebtToSS": {
                    "type": "boolean",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Social security benefits calculation"
          }
        }
      }
    },
    "/api/v1/simulate/iva-periodic": {
      "post": {
        "summary": "Simulate Periodic VAT Assessment & Quadro 06 Mapping",
        "description": "Calculates output VAT (23%, 13%, 6%), deductible VAT, Quadro 06 mapping, and net VAT payable vs. credit carried forward.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "grossInvoicedNormalRate": {
                    "type": "number",
                    "example": 15000
                  },
                  "grossInvoicedIntraEU": {
                    "type": "number",
                    "example": 5000
                  },
                  "deductibleVATEquipment": {
                    "type": "number",
                    "example": 600
                  },
                  "deductibleVATGeneralExpenses": {
                    "type": "number",
                    "example": 400
                  },
                  "priorPeriodVATCredit": {
                    "type": "number",
                    "example": 200
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Quarterly VAT periodic declaration result"
          }
        }
      }
    },
    "/api/v1/simulate/withholding-tax": {
      "post": {
        "summary": "Determine Category B Withholding Tax Rate (Art. 101º CIRS)",
        "description": "Calculates applicable withholding tax (25% Tabela 151, 16.5% other, 0% foreign/Art. 101-B), mandatory invoice clause, and net cash received.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "invoiceAmount": {
                    "type": "number",
                    "example": 2500
                  },
                  "serviceType": {
                    "type": "string",
                    "enum": [
                      "TABELA_151",
                      "OUTROS_SERVICOS",
                      "DIREITOS_AUTOR",
                      "COMISSOES"
                    ],
                    "example": "TABELA_151"
                  },
                  "isClientForeignEntity": {
                    "type": "boolean",
                    "example": false
                  },
                  "isClientParticularB2C": {
                    "type": "boolean",
                    "example": false
                  },
                  "optForArt101BExemption": {
                    "type": "boolean",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Withholding tax rate and clause"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key returned once by POST /api/v1/profiles, sent as 'Authorization: Bearer <apiKey>'."
      }
    }
  }
}